ingestion/requirements.txt
snowflake-connector-python==3.12.0
pandas==2.2.0
python-dotenv==1.0.0
